fix(sentry): unify dash+flow scalar decoding; verify selectability past a note failure; extract clear-failure CLI
Round 17 (#1769) — three P2s.

1. Decode quoted scalars in dash lists (project-core.mjs). The dash/block form
   (`- item`) only stripped outer quotes, while the inline flow form decoded the
   full escape set + single-quoted doubling — so `- "Yes → fix"` rendered
   the literal `→` and `- 'It''s fixed'` kept the doubled apostrophe. ROOT
   FIX: one shared `decodeScalar` routes a whole dash token through the SAME
   readers the flow parser uses, so a quoted dash item now decodes IDENTICALLY to
   a quoted flow item; a malformed quoted token keeps its content (outer quotes
   stripped), never corrupted. The escape set + `''` rule have ONE definition
   both list forms reach. The same-line scalar path uses it too.

2. Verify selectability even when the recovery note fails (requeue.mjs). Under
   verify-end-state, requeueQueueStub caught the label error but let a
   bookkeeping-NOTE post error escape before ensureSelectableForTriage ran, so a
   double GitHub failure could leave an open stub off sentry:needs-triage and
   report success. The note is advisory; the label restoration is load-bearing.
   The note post is now caught under verify-end-state (a notice), so end-state
   verification ALWAYS runs and a stub that cannot be confirmed selectable is a
   hard failure — never a success on an unverified strand. The code comment there
   anticipated exactly this caller.

3. Extract the clear-failure CLI (requeue.mjs 809 -> 663). The round-16 CLI is
   now scripts/sentry-triage-brief-clear-recovery.mjs (165 lines): the chokepoint
   stays pure dependency-injected logic, the sibling owns its I/O + argv shell.
   The workflow calls the new module; requeue.test.mjs imports the CLI from it;
   the file-size cap list, dependency-free-closure entries, and gate routing all
   include it. Both files under caps; closure stays third-party-free.

Self-review (round-16's CLI spawned 2 findings — not repeating): audited every
write path in requeueQueueStub before ensureSelectableForTriage (fence, label,
note) — all now caught under verify-end-state, no unverified escape; ingest's
abort-mode note still propagates (unchanged); decodeScalar's malformed /
unterminated / empty / colon edges are non-corrupting. Full brief + project +
requeue + archive + agent-comment + ingest + broker + CI-suites suites green.

Tests: a dash list with a double-quoted (escaped), single-quoted, and plain item
decodes the same as its flow equivalent; note-fails-but-labels-succeed is a
verified success; a label restoration that cannot be confirmed selectable is a
hard failure.
